Break apart the File class into multiple files and add type definitions.
