+++ .flowconfig
@@ -15,3 +15,4 @@ lib/third-party-libs.js.flow
 [options]
 suppress_comment= \\(.\\|\n\\)*\\$FlowFixMe
 suppress_comment= \\(.\\|\n\\)*\\$FlowIssue
+esproposal.export_star_as=enable
