+++ packages/babel-core/src/index.js
@@ -1,17 +1,20 @@
-export File from "./transformation/file";
-export buildExternalHelpers from "./tools/build-external-helpers";
+// @flow
+
+export {
+  default as buildExternalHelpers,
+} from "./tools/build-external-helpers";
 export { resolvePlugin, resolvePreset } from "./config/loading/files";
 
 export { version } from "../package";
 export { getEnv } from "./config/helpers/environment";
 
 export * as types from "babel-types";
-export traverse from "babel-traverse";
-export template from "babel-template";
+export { default as traverse } from "babel-traverse";
+export { default as template } from "babel-template";
 
 import loadConfig from "./config";
 
-export function loadOptions(opts): Object | null {
+export function loadOptions(opts: {}): Object | null {
   const config = loadConfig(opts);
 
   return config ? config.options : null;
@@ -19,21 +22,21 @@ export function loadOptions(opts): Object | null {
 
 // For easier backward-compatibility, provide an API like the one we exposed in Babel 6.
 export class OptionManager {
-  init(opts) {
+  init(opts: {}) {
     return loadOptions(opts);
   }
 }
 
-export function Plugin(alias) {
-  throw new Error(`The (${alias}) Babel 5 plugin is being run with Babel 6.`);
+export function Plugin(alias: string) {
+  throw new Error(
+    `The (${alias}) Babel 5 plugin is being run with an unsupported Babel version.`,
+  );
 }
 
-export {
-  transform,
-  transformFromAst,
-  transformFile,
-  transformFileSync,
-} from "./transformation";
+export { default as transform } from "./transform";
+export { default as transformFromAst } from "./transform-ast";
+export { default as transformFile } from "./transform-file";
+export { default as transformFileSync } from "./transform-file-sync";
 
 /**
  * Recommended set of compilable extensions. Not used in babel-core directly, but meant as
