+++ packages/babel-core/src/transform-file-sync.js
@@ -1,0 +1,16 @@
+// @flow
+import fs from "fs";
+
+import loadConfig from "./config";
+import runTransform, { type FileResult } from "./transformation";
+
+export default function transformFileSync(
+  filename: string,
+  opts?: Object = {},
+): FileResult | null {
+  opts.filename = filename;
+  const config = loadConfig(opts);
+  if (config === null) return null;
+
+  return runTransform(config, fs.readFileSync(filename, "utf8"));
+}
